requirements.txt
SQLAlchemy~=2.0.31
asyncmy~=0.2.9
Telethon~=1.36.0
cryptg~=0.4.0
Pillow~=10.3.0
aiohttp~=3.9.5
pycryptodome~=3.20.0
APScheduler~=3.10.4
jinja2~=3.1.4
aiosqlite~=0.20.0
